Fix splitbrain between `.js` being ESM & CJS inconsistently
